bugfix: Improve Key Store throughput
This bugfix is to improve key store throughput.
Performance is too bad due to the usage of poll(), and it even leads writer in hang.
The commit includes to use drainTo() instead of poll() that is implemented in LinkedBlockingQueue.
System load of ES will even decrease because no requests will be made for entries that are already have sent SearchRequest to ES before cache expires(default 5Min to expire).