bugfix: Prevent calling clearScroll if there's no scrollId.
Currently clearScroll() is called even if scrollIds is empty.
scrollIds is empty whenever there's no path in a target index.
If no documents are returned then the server should return 200 OK with empty payload.
By calling this method with an empty scrollIds, the response is 500 Internal Server error, which is not intended.
This bugfix prevent calling clearScroll() when scrollIds is empty.