bugfix: Prevent to set until parameter to current timestamp when until > current automatically.
We don't need to care about the until param because both ES and C* make empty response whenever there's no data corresponding to query.
Setting automatically in this manner causes another problem.
For example, let's consider the case that we are trying to compare the traffic between TODAY and the day before.
Let's say that the time of querying(current) is 12:00.
We're trying to get data from Graphene with following query:
target=a.b.c&target=timeShift(a,b,c, '1d')
If until is set to current timestamp then there's no problem.
But, if we want to query the data with time range between 00:00 ~ 23:59 then there's a problem.
Parameter until set to 12:00 so even timeShift function is also affected by the automatically changed time range.
Pretty weird situation because we cannot get YESTERDAY's data ahead of 12:00 ~.
This commit removes automatic change of time range.

Plus, purged unnecessary compile jobs that are found in module's build.gradle.
Plus plus, changed not to call query in CassandraMetricService.java. This is unnecessary also.