bugfix: Fixed bug in SumSeriesWithWildcards and refactored TimeShift related codes.
SumSeriesWithWildcards doesn't work twice because it checks whether a given argument is PathTarget or not. This is nonsense because one can try to apply the function one more time. I've fixed it.

Seems there's a confuse to add minus sign to offset and subtract it in shift() method.
Offset should be added no matter what the sign is.
Changed the philosophy of method and refactored the code for readability and null safety.